Balance pass: gate win, stacking, long money model, boss/shop fixes
Quick wins:
- $1M no longer auto-wins mid-run; YouWin only after the final (stage 8)
  boss is cleared, so bosses are the real gates and a big overshoot just
  banks money.
- Clamp tooltips to the window (items near the right/bottom edge no longer
  clip off-screen).
- The Devil's second debuff is now DestroyValuableSegmentEffect (eats the
  most valuable segment before each spin) instead of a Crow-like red ban.

Economy:
- Explicit sub-linear stacking for Paint It Black / Scarlet Surge via a
  primary-copy pattern (Inventory.countFortunes/isPrimaryFortune); other
  fortunes stack as independent listeners. Duplicates allowed again.
- Interest reworked to flat $10*stage + 3% balance (front-loaded; weak late).
- First shop 40% cheaper; Ricochet re-rolls any losing tile; Lucky Seven
  triggers on 7/17/27.

Bug fixes:
- Money model int -> long (Player balance, payout path, MathHelper) to stop
  a huge win overflowing into a negative balance / instant loss. Bet
  placement stays int-bounded (fine for the campaign; endless TODO).
- Boss goal no longer pre-credited with the stage-clearing spin's interest
  (snapshot moved to beginBossFight).
- Shop renderer refreshed after a boss reward so the next stage's stock no
  longer shows the previous shop's sold-out items.

+++ core/src/main/java/de/mario/roguelette/Player.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class Player {
     private final Inventory inventory;
 
-    private int balance;
+    private long balance;
     private int currentlyInHand = 0; //amount of money currently in hand to be placed somewhere
     private final String name;
 
@@ -36,7 +36,7 @@ public class Player {
         return characterListeners;
     }
 
-    public int getBalance() {
+    public long getBalance() {
         return balance;
     }
 
@@ -60,15 +60,15 @@ public class Player {
         return inventory;
     }
 
-    public void pay(int amount) {
+    public void pay(long amount) {
         this.balance -= amount;
     }
 
-    public void earn(int amount) {
+    public void earn(long amount) {
         this.balance += amount;
     }
 
-    public boolean canAfford(int amount) {
+    public boolean canAfford(long amount) {
         // strictly greater as a balance of 0 means game over
         return this.balance > amount;
     }
